+++ quint/src/effects/ToEffectVisitor.ts
@@ -12,7 +12,7 @@
  * @module
  */
 
-import { Effect, Variables, emptyVariables } from './base'
+import { Effect, Variables } from './base'
 import { EffectListener } from '../generated/EffectListener'
 import * as p from '../generated/EffectParser'
 
@@ -28,6 +28,7 @@ export class ToEffectVisitor implements EffectListener {
   private arrowEffectsStack: Effect[][] = []
   private variablesStack: Variables[] = []
   private stateVars: string[] = []
+  private idCounter: bigint = 0n
 
   private pushEffect(effect: Effect): void {
     this.effect = effect
@@ -39,19 +40,19 @@ export class ToEffectVisitor implements EffectListener {
 
   exitReadOnly() {
     const variables = this.variablesStack.pop()!
-    const effect: Effect = { kind: 'concrete', read: variables, update: emptyVariables, temporal: emptyVariables }
+    const effect: Effect = { kind: 'concrete', components: [{ kind: 'read', variables }] }
     this.pushEffect(effect)
   }
 
   exitUpdateOnly() {
     const variables = this.variablesStack.pop()!
-    const effect: Effect = { kind: 'concrete', read: emptyVariables, update: variables, temporal: emptyVariables }
+    const effect: Effect = { kind: 'concrete', components: [{ kind: 'update', variables }] }
     this.pushEffect(effect)
   }
 
   exitTemporalOnly() {
     const variables = this.variablesStack.pop()!
-    const effect: Effect = { kind: 'concrete', read: emptyVariables, update: emptyVariables, temporal: variables }
+    const effect: Effect = { kind: 'concrete', components: [{ kind: 'temporal', variables }] }
     this.pushEffect(effect)
   }
 
@@ -59,7 +60,10 @@ export class ToEffectVisitor implements EffectListener {
     const update = this.variablesStack.pop()!
     const read = this.variablesStack.pop()!
 
-    const effect: Effect = { kind: 'concrete', read, update, temporal: emptyVariables }
+    const effect: Effect = {
+      kind: 'concrete',
+      components: [{ kind: 'read', variables: read }, { kind: 'update', variables: update }],
+    }
     this.pushEffect(effect)
   }
 
@@ -67,12 +71,15 @@ export class ToEffectVisitor implements EffectListener {
     const temporal = this.variablesStack.pop()!
     const read = this.variablesStack.pop()!
 
-    const effect: Effect = { kind: 'concrete', read, update: emptyVariables, temporal }
+    const effect: Effect = {
+      kind: 'concrete',
+      components: [{ kind: 'read', variables: read }, { kind: 'temporal', variables: temporal }],
+    }
     this.pushEffect(effect)
   }
 
   exitPure() {
-    const effect: Effect = { kind: 'concrete', read: emptyVariables, update: emptyVariables, temporal: emptyVariables }
+    const effect: Effect = { kind: 'concrete', components: [] }
     this.pushEffect(effect)
   }
 
@@ -97,7 +104,7 @@ export class ToEffectVisitor implements EffectListener {
     const names: string[] = ctx.IDENTIFIER().map(i => i.text)
     const unionVariables: Variables[] = names.map(name => ({ kind: 'quantified', name }))
     if (this.stateVars.length > 0) {
-      unionVariables.push({ kind: 'concrete', vars: this.stateVars })
+      unionVariables.push({ kind: 'concrete', vars: this.stateVars.map(v => ({ name: v, reference: this.nextId() })) })
     }
 
     if (unionVariables.length === 0) {
@@ -114,5 +121,9 @@ export class ToEffectVisitor implements EffectListener {
   exitStateVarRef(ctx: p.StateVarRefContext) {
     const varRef = ctx.IDENTIFIER().text
     this.stateVars.push(varRef)
+  }
+
+  private nextId(): bigint {
+    return ++this.idCounter
   }
 }