+++ vscode/quint-vscode/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to the Quint VSCode extension will be documented in this fil
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+### Added
+
+- Support for "Go to definition" (#546).
+
 ## v0.1.1 -- 2023-01-17
 
 ### Added