+++ vscode/quint-vscode/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  - Change parsing interfaces to support multiple modules (#547)
  - Change static analysis interfaces to support multiple modules (#559)
+ - Change documentation overlays and code navigation to support multiple modules (#584)
 
 ## v0.1.2 -- 2023-01-24