+++ vscode/quint-vscode/CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## UNRELEASED
+        
+        ### Added
+        ### Changed
+        ### Deprecated
+        ### Removed
+        ### Fixed
+        ### Security
+        
+        ## v0.1.2 -- 2023-02-22
 
 ### Added